12.12.2023

See merge request foxstudent104063/school-console-app!15

+++ schoolconsoleapp/src/main/java/ua/com/foxminded/yuriy/schoolconsoleapp/service/GroupService.java
@@ -5,7 +5,7 @@ import ua.com.foxminded.yuriy.schoolconsoleapp.entity.Group;
 
 public interface GroupService {
 
-	List<Group> getAllLessOrEqual(int studentsCount);
+	List<Group> getAllLessOrEqual(Long studentsCount);
 
 	List<Group> getAll();
